requirements.txt
esptool==4.8.0
pyserial==3.5
